DEV: Pickups stats adjustements made by persent of the stat maximum

+++ TheFirmaTheBetter/Assets/Project/Scripts/ShipSelection/ShipParts/ShipResources.cs
@@ -92,7 +92,7 @@ namespace ShipParts
             Channels.OnPlayerStatsChanged?.Invoke(shipBuilder, shipStats);
         }
 
-        public int CurrentEnergyAmount => shipEnergy.CurrentEnergyAmount;
+        public float CurrentEnergyAmount => shipEnergy.CurrentEnergyAmount;
 
         public ShipStats ShipStats => shipStats;
